improve summary page + refactor
