+++ src/react-challenges/Challenges.tsx
@@ -1,89 +1,31 @@
 import { useState } from "react";
 import { Challenge, MappingsUsed } from "./Terminal";
-
-const challenges: Challenge[] = [
-  {
-    strokes: 1,
-    description: "Delete line 1",
-    content: "Once upon a time\nThere was a little pig named Percy\nPercy loved adventures\nOne day, he found a map",
-    expected: "There was a little pig named Percy\nPercy loved adventures\nOne day, he found a map"
-  },
-  {
-    strokes: 1,
-    description: "Navigate to end of line 1",
-    content: "Percy lived on a farm\nHe had many friends\nThey often played together\nLife was joyful",
-    expected: (vim) => vim.cursorPos.startLine == 0 && vim.cursorPos.startIndex == vim.content[0].length - 1
-  },
-  {
-    strokes: 1,
-    description: "Enter insert mode",
-    content: "Percy loved to explore\nEvery day was a new adventure\nHe discovered hidden paths\nAnd secret treasures",
-    expected: (vim) => vim.mode === "Insert"
-  },
-  {
-    strokes: 1,
-    description: "Enter visual mode",
-    content: "On a sunny day\nPercy found an old map\nIt led to a hidden treasure\nHis heart raced with excitement",
-    expected: (vim) => vim.mode === "Visual"
-  },
-  {
-    strokes: 2,
-    description: "Copy line 3",
-    content: "The map was ancient\nIt showed a forest\nDeep within the forest\nWas a treasure marked with an X",
-    expected: (vim) => {
-      return vim.clipboard.content === "Deep within the forest\n"
-    }
-  },
-  {
-    strokes: 1,
-    description: "Create line above line 2",
-    content: "Percy gathered his friends\nThey were excited\nTogether they set off\nOn a grand adventure",
-    expected: (vim) => {
-      return vim.content.length == 5 && vim.content[1] === " ";
-    }
-  },
-  {
-    strokes: 1,
-    description: "Go to line 2",
-    content: "The journey was long\nBut they were determined\nThey faced many challenges\nAnd overcame them all",
-    expected: (vim) => {
-      return vim.cursorPos.startLine === 1;
-    }
-  },
-  {
-    strokes: 1,
-    description: "Move cursor to the right",
-    content: "At last, they arrived\nThe treasure was real\nThey couldn't believe their eyes\nIt was a dream come true",
-    expected: (vim) => {
-      return vim.cursorPos.startIndex === 1
-    }
-  },
-  {
-    strokes: 1,
-    description: "Navigate to end of buffer",
-    content: "With the treasure found\nThey returned home\nTheir adventure was complete\nBut memories would last forever",
-    expected: (vim) => vim.cursorPos.startLine === vim.content.length - 1
-  },
-  {
-    strokes: 2,
-    description: "Navigate to end of line 3",
-    content: "Percy was a brave pig\nHe loved discovering new places\nHis friends admired his courage\nThey often followed him on adventures",
-    expected: (vim) => vim.cursorPos.startLine == 2 && vim.cursorPos.startIndex == vim.content[2].length - 1
-  }
-]
+import { Button } from "./components/Button";
+import { challenges } from './challenges'
+import React from "react";
 
 export function Challenges() {
   const [challengeIndex, setChallengeIndex] = useState(0)
+  const isFinished = challengeIndex == challenges.length
   const [mappingsUsed, setMappingsUsed] = useState(0);
   const challenge = challenges[challengeIndex]
-  if (!challenge) {
+
+  if (isFinished) {
     const leastPossibleMappings = challenges.reduce((acc, curr) => acc + curr.strokes, 0 as number)
     return (
-      <div>
+      <div style={{ display: "flex", flexDirection: "column", gap: "8px" }}>
         <h2>Finished</h2>
-        <MappingsUsed least={leastPossibleMappings} spent={mappingsUsed}/>
+        <MappingsUsed least={leastPossibleMappings} spent={mappingsUsed} />
+        <div>
+          <Button>Try again</Button>
+        </div>
       </div>)
   }
+
+  if (!challenge) {
+    return <div>Something went wrong</div>
+  }
+
   return (
     <div>
       <div style={{ height: "20px" }}>{challengeIndex + 1}/{challenges.length}</div>
